Optimize debug_traceTransaction for memory usage
Optimize debug_traceTransaction for memory usage by using objects to track data within the trace. Most of the optimization occurs in the TraceData class itself, where values are memoized and only create a new TraceData object if one isn't created for that value yet. In the associated test, this reduces the number of objects used to create the trace by 47%.